+++ bluetooth_low_energy_platform_interface/lib/src/my_peripheral.dart
@@ -5,4 +5,12 @@ class MyPeripheral extends MyBluetoothLowEnergyPeer implements Peripheral {
   MyPeripheral({
     required super.uuid,
   });
+
+  @override
+  int get hashCode => uuid.hashCode;
+
+  @override
+  bool operator ==(Object other) {
+    return other is Peripheral && other.uuid == uuid;
+  }
 }